Timestamps are stored as integer values

+++ libs/lensfun/auxfun.cpp
@@ -10,6 +10,7 @@
 #include <ctype.h>
 #include <math.h>
 #include <fstream>
+#include <stdint.h>
 
 static const char *_lf_get_lang ()
 {
@@ -461,9 +462,9 @@ float _lf_interpolate (float y1, float y2, float y3, float y4, float t)
         (t3 - t2) * tg3;
 }
 
-float _lf_read_database_timestamp (const gchar *dirname)
+long int _lf_read_database_timestamp (const gchar *dirname)
 {
-    float timestamp = -1.0;
+    long int timestamp = -1;
     GDir *dir = g_dir_open (dirname, 0, NULL);
     if (dir)
     {
